+++ app/visualize/analysis/stmt/parser/for_stmt.py
@@ -1,7 +1,7 @@
 import ast
 
 from app.visualize.container.element_container import ElementContainer
-from app.visualize.analysis.stmt.expr.expr_traveler import ExprTraveler
+from app.visualize.analysis.stmt.parser.expr.expr_traveler import ExprTraveler
 from app.visualize.analysis.stmt.models.for_stmt_obj import ForStmtObj
