+++ tests/visualize/analysis/stmt/expr/parser/test_call_expr.py
@@ -1,8 +1,8 @@
 import pytest
 
-from app.visualize.analysis.stmt.expr.models.expr_obj import ExprObj, ConstantObj, NameObj, BinopObj
-from app.visualize.analysis.stmt.expr.models.range_expression import RangeExpression
-from app.visualize.analysis.stmt.expr.parser.call_expr import CallExpr
+from app.visualize.analysis.stmt.parser.expr.models.expr_obj import ExprObj, ConstantObj, NameObj, BinopObj
+from app.visualize.analysis.stmt.parser.expr.models.range_expression import RangeExpression
+from app.visualize.analysis.stmt.parser.expr.parser.call_expr import CallExpr
 
 
 @pytest.mark.parametrize("func_name, args, keyword_arg_dict, expected", [])
