+++ tests/visualize/analysis/stmt/expr/parser/test_compare_expr.py
@@ -2,8 +2,8 @@ import ast
 
 import pytest
 
-from app.visualize.analysis.stmt.expr.models.expr_obj import CompareObj, ConstantObj, NameObj
-from app.visualize.analysis.stmt.expr.parser.compare_expr import CompareExpr
+from app.visualize.analysis.stmt.parser.expr.models.expr_obj import CompareObj, ConstantObj, NameObj
+from app.visualize.analysis.stmt.parser.expr.parser.compare_expr import CompareExpr
 
 
 @pytest.mark.parametrize(
